+++ rawtools/nsihdr.py
@@ -3,16 +3,20 @@
 """NSIHDR to RAW Batch Converter"""
 import logging
 import os
-import re
 import sys
+import tkinter as tk
 from multiprocessing import Pool, cpu_count
+from pprint import pformat
 from time import time
-
+from tkinter import (E, N, S, StringVar, Toplevel, W, filedialog, messagebox,
+                     ttk)
+import threading
 import numpy as np
 from tqdm import tqdm
 
 from rawtools import dat
 from rawtools.convert import scale
+from rawtools.gui import nsihdr
 
 # Load in NSI SDK
 currentdir = os.path.dirname(os.path.realpath(__file__))
@@ -21,104 +25,47 @@ includesdir = os.path.join(rootdir, "bin")
 sys.path.append(includesdir)
 from rawtools import nsiefx
 
+# def check_progress(amount):
+# 	logging.info(amount)
+# 	# progress_text.set(str(round(total_slices_processed / args.total_slice_count * 100.0, 1)))
+# 	progress_text.set(str(total_slices_processed))
+# 	progress['value'] = total_slices_processed
+# 	logging.info(f"{progress_text.get()=}")
 
-def write_metadata(args, metadata):
-	"""Generates a .dat file from information gathered from an .nsihdr file
 
-	NOTE(tparker): Temporarily, I am writing the minimum and maximum values found
-	in the 32-bit float version of the files in case we ever need to convert the
-	uint16 version back to float32.
+def update_progress(increment):
+	logging.debug(f"{increment=}")
 
-	Args:
-		args (ArgumentParser): user arguments from `argparse`
-		metadata (dict): dictionary of metadata created from reading .nsihdr file
-	"""
-	ObjectFileName = args.output
-	resolution = ' '.join(metadata['dimensions'])
-	slice_thickness = ' '.join([ str(rr) for rr in metadata['resolution_rounded'] ])
-	dat_filepath = f'{os.path.splitext(args.output)[0]}.dat'
-	output_string = f"""ObjectFileName: {ObjectFileName}\nResolution:     {resolution}\nSliceThickness: {slice_thickness}\nFormat:         {metadata['bit_depth_type']}\nObjectModel:    {metadata['ObjectModel']}"""
+def start_progress_thread(event, pbar, root):
+	global progress_thread
+	progress_thread = threading.Thread(target=update_progress, args=(1,))
+	progress_thread.daemon = True
+	# pbar.start()
+	progress_thread.start()
+	root.after(20, check_progress_thread(pbar, root))
 
-	dat.write(dat_filepath, metadata['dimensions'], metadata['resolution_rounded'])
-	# with open(dat_filepath, 'w') as ofp:
-	#   print(f'Generating {dat_filepath}')
-	#   ofp.write(output_string)
+def check_progress_thread(pbar, root):
+	if progress_thread.is_alive():
+		logging.info("Pbar is alive")
+		root.after(20, check_progress_thread)
+	else:
+		logging.info("Pbar is done")
+		pbar.stop()
 
-	bounds_filepath = os.path.join(args.cwd, f'{os.path.splitext(args.output)[0]}.float32.range')
-	with open(bounds_filepath, 'w') as ofp:
-		print(f'Generating {bounds_filepath}')
-		bounds = f'{INITIAL_LOWER_BOUND} {INITIAL_UPPER_BOUND}'
-		ofp.write(bounds)
 
-def read_nsihdr(args, fp):
-	"""Collects relative metadata from .nsihdr file
-
-	Args:
-		fp (str): Input filepath to an .nsihdr file
-
-	Returns:
-		dict: metadata about NSI project
-	"""
-	global INITIAL_LOWER_BOUND
-	global INITIAL_UPPER_BOUND
-
-	with open(fp, 'r') as ifp:
-		document = ifp.readlines()
-
-		source_to_detector_distance = None
-		source_to_table_distance = None
-		bit_depth = None
-		nsidats = []
-
-		bit_depth_query = re.search(bit_depth_pattern, line)
-		if bit_depth_query:
-			bit_depth = int(bit_depth_query.group('value'))
-
-		dimensions_query = re.search(dimensions_pattern, line)
-		if dimensions_query:
-			dimensions = [ dimensions_query.group('x'), dimensions_query.group('z'), dimensions_query.group('num_slices') ]
-
-		# Check if the .nsihdr already contains the data range
-		# If it exists, we only have to read the .nsidat files once instead of twice
-		data_range_query = re.search(data_range_pattern, line)
-		if data_range_query:
-			INITIAL_LOWER_BOUND = float(data_range_query.group('lower_bound'))
-			INITIAL_UPPER_BOUND = float(data_range_query.group('upper_bound'))
-
-		# Temporarily set pitch as 0.127, as it should not change until we get a
-		# new detector
-		pitch = 0.127
-
-		# TODO(tparker): As far as I am aware, the data will always be of type DENSITY
-		ObjectModel = 'DENSITY'
-
-		resolution = ( pitch / source_to_detector_distance ) * source_to_table_distance
-		resolution_rounded = round(resolution, 4)
-		nsidats.sort() # make sure that the files are in alphanumeric order
-
-		return {
-			"datafiles": nsidats,
-			"source_to_detector_distance": source_to_detector_distance,
-			"source_to_table_distance": source_to_table_distance,
-			"pitch": pitch,
-			"resolution": resolution,
-			"resolution_rounded": [resolution_rounded]*3,
-			"bit_depth": bit_depth,
-			"zoom_factor": round(source_to_detector_distance / source_to_table_distance, 2),
-			"bit_depth_type": dat.bitdepth(bit_depth),
-			"ObjectModel": ObjectModel,
-			"dimensions": dimensions
-		}
-
-def process(args, fp, ofp):
+def process(args, fp, export_path):
 	"""Converts NSIHDR files to a single .RAW + .DAT
 
 	Args:
+
 		args (ArgumentParser): user arguments from `argparse`
 		fp (str): filepath to input .NSIHDR file
-		ofp (str): filepath to output .RAW file
+		export_path (str): filepath to output .RAW file
 	"""
 	logging.debug(f'{fp=}')
+	total_slices_processed = 0
+	progress = None
+	check_progress = None
 
 	with nsiefx.open(fp) as volume:
 		v = volume # for shorthand laziness
@@ -137,12 +84,7 @@ def process(args, fp, ofp):
 
 		dname = os.path.dirname(fp)
 		bname = os.path.basename(os.path.splitext(fp)[0])
-		export_path = os.path.join(dname, f'{bname}.raw')
-		logging.debug(f"{export_path=}")
 		dat_path = os.path.join(dname, f'{bname}.dat')
-		logging.debug(f"{dat_path=}")
-
-		dat.write(dat_path, dimensions = (width, height, depth), thickness = voxel_size)
 
 		if os.path.exists(export_path) and args.force == True:
 			os.remove(export_path)
@@ -151,15 +93,26 @@ def process(args, fp, ofp):
 			os.remove(dat_path)
 			logging.warning(f"Removed old '{dat_path}'")
 
+		dat.write(dat_path, dimensions = (width, height, depth), thickness = voxel_size)
+		logging.debug(f"Generated '{dat_path}'")
+
+		pbar = None
 		with open(export_path, 'ab') as raw_ofp:
-			pbar = tqdm(total= depth, desc=f"Exporting {bname}")
+			if not args.verbose:
+				pbar = tqdm(total= depth, desc=f"Exporting {bname}")
 			for n in range(depth):
 				cross_section = v.read_slice(n)
 				cross_section = np.array(cross_section, dtype="float32")
 				cross_section = scale(cross_section, data_min, data_max, 0, 65535).astype(np.uint16)
 				cross_section.tofile(raw_ofp)
-				pbar.update()
-			pbar.close()
+
+				total_slices_processed += 1
+
+				if not args.verbose:
+					pbar.update()
+				# logging.debug(f"Processed {total_slices_processed}")
+			if not args.verbose:
+				pbar.close()
 
 def main(args):
 	start_time = time()
@@ -168,48 +121,134 @@ def main(args):
 		# Gather all files
 		args.files = []
 		for p in args.path:
-			for root, dirs, files in os.walk(p):
+			for root, _, files in os.walk(p):
 				for filename in files:
 					args.files.append(os.path.join(root, filename))
 
-		# Append any loose, explicitly defined paths to .RAW files
+		# Append any loose, explicitly defined paths to .nsihdr files
 		args.files.extend([ f for f in args.path if f.endswith('.nsihdr') ])
 
 		# Filter out non-NSIHDR files
 		args.files = [ f for f in args.files if f.endswith('.nsihdr') ]
 
 		# Get all RAW files
-		logging.debug(f"All files: {args.files}")
+		logging.debug(f"All files: {pformat(args.files)}")
 		args.files = list(set(args.files)) # remove duplicates
-		logging.info(f"Found {len(args.files)} volume(s).")
-		logging.debug(f"Unique files: {args.files}")
+		logging.debug(f"Unique files: {pformat(args.files)}")
+
+		# If file overwriting is disabled
+		if not args.force:
+			kept_volumes = []
+			skipped_volumes = []
+			for fp in args.files:
+				dname = os.path.dirname(fp)
+				bname = os.path.basename(os.path.splitext(fp)[0])
+				export_path = os.path.join(dname, f'{bname}.raw')
+				if os.path.exists(export_path) and os.path.isfile(export_path):
+					skipped_volumes.append(fp)
+				else:
+					kept_volumes.append(fp)
+			args.files = kept_volumes
+			total_volumes = len(kept_volumes) + len(skipped_volumes)
+			logging.debug(f"{kept_volumes=}")
+			logging.debug(f"{skipped_volumes=}")
+
+			logging.info(f"Found {total_volumes} volume(s). (Unchanged: {len(kept_volumes)}, Skipped: {len(skipped_volumes)})")
+
+		# Otherwise, overwrite files
+		else:
+			unprocessed_volumes = []
+			existing_volumes = []
+			for fp in args.files:
+				dname = os.path.dirname(fp)
+				bname = os.path.basename(os.path.splitext(fp)[0])
+				export_path = os.path.join(dname, f'{bname}.raw')
+				if os.path.exists(export_path) and os.path.isfile(export_path):
+					existing_volumes.append(export_path)
+				else:
+					unprocessed_volumes.append(export_path)
+			total_volumes = len(existing_volumes) + len(unprocessed_volumes)
+
+			logging.debug(f"{existing_volumes=}")
+			logging.debug(f"{unprocessed_volumes=}")
+
+			logging.info(f"Found {total_volumes} volume(s). (Overwriting: {len(existing_volumes)}, New: {len(unprocessed_volumes)})")
+
 	except Exception as err:
 		logging.error(err)
 		raise err
 	else:
+
+		# GUI Implementation
+		if args.gui:
+			# Determine the number of slices in advance
+			args.total_slice_count = 0
+			for fp in args.files:
+				with nsiefx.open(fp) as volume:
+					args.total_slice_count += volume.num_slices()
+			
+			# Initialize progress bar
+			progress_bar_prompt_title = "Placeholder Progress Bar"
+			app = args.app
+			root = app.root
+			icon_fp = app.icon_fp
+			progress_bar_prompt = Toplevel(root)
+			progress_bar_prompt.title(progress_bar_prompt_title)
+			progress_bar_prompt.iconbitmap(icon_fp)
+			progress_bar_prompt.resizable(False, False)
+			progress_bar_prompt_frame = ttk.Frame(progress_bar_prompt, padding="16 16")
+			progress_bar_prompt_frame.grid(column=0, row=0, sticky=(N, S, E, W))
+
+			progress = ttk.Progressbar(progress_bar_prompt_frame, orient='horizontal', mode='indeterminate', length=200, max=args.total_slice_count)
+			progress.grid(row=0, column=0, columnspan=2, pady="0 16", sticky=(E,W))
+			progress_text = tk.StringVar()
+			progress_text_label = ttk.Label(progress_bar_prompt_frame, textvariable=progress_text, width=7)
+			progress_text_label.grid(row=0, column=3, columnspan=1, pady="0 16", sticky=E, padx="8 0")
+			progress_text.set('0%')
+
+			def dismiss_progress_prompt():
+				progress_bar_prompt.grab_release()
+				progress_bar_prompt.destroy()
+
+			# Orient window on screen
+			nsihdr.center(root, progress_bar_prompt)
+			# Disable interaction with parent window
+			progress_bar_prompt.protocol("WM_DELETE_WINDOW", dismiss_progress_prompt)
+			progress_bar_prompt.transient(root)
+			progress_bar_prompt.wait_visibility()
+			progress_bar_prompt.grab_set()
+			progress_bar_prompt.wait_window()
+			start_progress_thread(None, progress, root)
+			
+		# CLI Implementation
 		# For each provided volume...
-		pbar = tqdm(total = len(args.files), desc=f"Overall progress")
+		pbar = None
+		if not args.verbose:
+			pbar = tqdm(total = len(args.files), desc=f"Overall progress")
+
 		for fp in args.files:
 			logging.debug(f"Processing '{fp}'")
-			ofp_directory = os.path.dirname(fp)
-			logging.debug(f"{ofp_directory=}")
-			ofp_filename = os.path.basename(os.path.splitext(fp)[0])
-			logging.debug(f"{ofp_filename=}")
-			ofp = os.path.join(ofp_directory, ofp_filename)
-			logging.debug(f"{ofp=}")
+			dname = os.path.dirname(fp)
+			bname = os.path.basename(os.path.splitext(fp)[0])
+			export_path = os.path.join(dname, f'{bname}.raw')
+			logging.debug(f"{export_path=}")
+			dat_path = os.path.join(dname, f'{bname}.dat')
+			logging.debug(f"{dat_path=}")
 
 			# Determine output location and check for conflicts
-			if os.path.exists(ofp) and os.path.isfile(ofp):
+			if os.path.exists(export_path) and os.path.isfile(export_path):
 				# If file creation not forced, do not process volume, return
 				if args.force == False:
-					logging.info(f"File already exists. Skipping {ofp}.")
+					logging.info(f"File already exists. Skipping {export_path}.")
 					continue
 				# Otherwise, user forced file generation
 				else:
-					logging.warning(f"FileExistsWarning - {ofp}. File will be overwritten.")
+					logging.warning(f"FileExistsWarning - {export_path}. File will be overwritten.")
 
 			# Extract slices and cast to desired datatype
-			process(args, fp, ofp)
+			process(args, fp, export_path)
 
-			pbar.update()
-		pbar.close()
+			if not args.verbose:
+				pbar.update()
+		if not args.verbose:
+			pbar.close()